variation with random room assignment
made a random model of staff assignment to rooms in which the minimum number of staff assigned to each room on each day and the minimum number of rooms assigned to a staff member on each day are the only constraints other than the roster determining which workers are present on a given day. Also did away with the 'network generator'  directory because this part is now done in the Julia script that runs the model.