changed immunity and transmission scalar
adjusted to a generic immunity distribution corresponding to the bimodal efficacy paper parameters for the ancestral variant from Khoury et al. and Cromer et al. Also changed the base implementation to reflect some of the useful changes in the R0 implementation. Still need to change the way the population is initialised (resetting agents instead of reinitialising, randomising test schedule, etc.). Should also consider randomising the roster for each run, but would have to consider if the development time is justified. Cleaned up to remove mention of 'delta variant' or 'omicron variant' etc.