corrected hom version for contact scaling
contact numbers scale with number of residents in structured model, but do not depend on the number of staff members. This is now accounted for in the homogeneous approximation which must have equivalent numbers of contacts for the comparison to be valid. The correction factor applied was 0.625 accounting for the number of person-days contributed by staff on intermittent roster (FTE * 5/7).