updated analysis with network isolation QC results
checked the extent to which residents get isolated from service due to furlough of staff members - in the worst-case scenario this happens in about 10% of outbreaks, and in over half of those the number of resident-days of isolation is 1 (i.e., one resident cut off from staff for one day over the whole simulation). The highest number recorded was 9 resident-days, and this occurred in only one instance (1/1000).